pyproject.toml
[tool.poetry]
name = "heating-planner"
version = "0.1.0"
description = ""
authors = ["François Weber <francois.weber.commercial@gmx.fr>"]
readme = "README.md"
packages = [{ include = "heating_planner" }]

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.0.3"
numpy = "^1.25.2"
seaborn = "^0.12.2"
tabulate = "^0.9.0"
ipywidgets = "^8.1.0"
typer = "^0.9.0"
loguru = "^0.7.0"
pygal-maps-fr = "^1.1.0"
cairosvg = "^2.7.1"
streamlit = "^1.25.0"
geopy = "^2.3.0"
tqdm = "^4.66.1"
plotly = "^5.16.1"
scikit-learn = "^1.3.0"
samplerate = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
